Separate components into new files.
